// backend/car-rental/src/modules/cars/models/create-car.model.ts
import { CarTypeEnum } from '../enums/car-type.enum';
import { CarDriveEnum } from '../enums/car-drive.enum';

export interface CreateCarParameters {
  horsePower: number;
  maxSpeed: number;
  hundredSpeed: number;
  engineCapacity: number;
  drive: CarDriveEnum;
}

export interface CreateCar {
  brand: string;
  model: string;
  year: string;
  description?: string;
  type: CarTypeEnum;
  image: string;
  carParameters: CreateCarParameters;
}